refactor: Shop 필드값 수정

+++ src/main/java/in/koreatech/koin/domain/shop/dto/ModifyShopRequest.java
@@ -24,7 +24,7 @@ public record ModifyShopRequest(
     Boolean delivery,
 
     @Schema(example = "1000", description = "배달비")
-    Long deliveryPrice,
+    Integer deliveryPrice,
 
     @Schema(example = "string", description = "설명")
     String description,
