+++ src/main/java/in/koreatech/koin/domain/shop/dto/ShopResponse.java
@@ -24,7 +24,7 @@ public record ShopResponse(
     Boolean delivery,
 
     @Schema(example = "1000", description = "배달비")
-    Long deliveryPrice,
+    Integer deliveryPrice,
 
     @Schema(example = "string", description = "설명")
     String description,
